requirements.txt
asgiref==3.3.4
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.6
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.3
django-bootstrap-datepicker-plus==3.0.5
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-heroku==0.3.1
django-import-export==2.5.0
django-jazzmin==2.4.7
django-js-asset==1.2.2
django-prometheus==2.1.0
django-widget-tweaks==1.4.8
et-xmlfile==1.1.0
flake8==3.9.2
gunicorn==20.1.0
html5lib==1.1
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
prometheus-client==0.11.0
#==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyphen==0.11.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
tinycss2==1.1.0
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
